Prevent oauth user from editing email or password fields

+++ inventory/routes.py
@@ -13,7 +13,7 @@ from inventory import app, bcrypt, db, mail
 from inventory.form import (AssetForm, DeleteAccountForm, LoginForm,
                             PasswordResetForm, RegisterForm, RequestResetForm,
                             UpdateAccountForm, UpdateAssetForm,
-                            UpdatePasswordForm)
+                            UpdateOauthAccountForm, UpdatePasswordForm)
 from inventory.models import Asset, User
 
 unit_options = {
@@ -198,31 +198,46 @@ def get_units():
 @app.route("/profile/", methods=["GET", "POST"])
 @login_required
 def profile():
+    oauth_form = UpdateOauthAccountForm()
     form = UpdateAccountForm()
     password_form = UpdatePasswordForm()
-    if request.method == "GET":
-        form.full_name.data = current_user.full_name
-        form.email.data = current_user.email
-    if form.validate_on_submit():
-        if (
-            current_user.full_name == form.full_name.data
-            and current_user.email == form.email.data
-        ):
-            redirect(url_for("profile"))
-        else:
-            print(current_user.full_name)
-            current_user.full_name = form.full_name.data.title()
-            current_user.email = form.email.data.lower()
+    if current_user.is_oauth_user == True:
+        if request.method == "GET":
+            oauth_form.full_name.data = current_user.full_name
+        if oauth_form.validate_on_submit():
+            if (
+                current_user.full_name == oauth_form.full_name.data
+            ):
+                redirect(url_for("profile"))
+            else:
+                current_user.full_name = oauth_form.full_name.data.title()
+                db.session.commit()
+                flash("Your Account Info has been updated successfully!", "success")
+                return redirect(url_for("profile"))
+    else:
+        if request.method == "GET":
+            form.full_name.data = current_user.full_name
+            form.email.data = current_user.email
+        if form.validate_on_submit():
+            if (
+                current_user.full_name == form.full_name.data
+                and current_user.email == form.email.data
+            ):
+                redirect(url_for("profile"))
+            else:
+                print(current_user.full_name)
+                current_user.full_name = form.full_name.data.title()
+                current_user.email = form.email.data.lower()
+                db.session.commit()
+                flash("Your Account Info has been updated successfully!", "success")
+                return redirect(url_for("profile"))
+        if password_form.validate_on_submit():
+            hashed_password = bcrypt.generate_password_hash(password_form.new_password.data)
+            current_user.password = hashed_password
             db.session.commit()
-            flash("Your Account Info has been updated successfully!", "success")
+            flash("Your Password has been updated successfully!", "success")
             return redirect(url_for("profile"))
-    if password_form.validate_on_submit():
-        hashed_password = bcrypt.generate_password_hash(password_form.new_password.data)
-        current_user.password = hashed_password
-        db.session.commit()
-        flash("Your Password has been updated successfully!", "success")
-        return redirect(url_for("profile"))
-    return render_template("user_profile.html", form=form, password_form=password_form)
+    return render_template("user_profile.html", form=form, password_form=password_form,oauth_form=oauth_form)
 
 
 @app.route("/signin/", methods=("GET", "POST"))
@@ -297,6 +312,8 @@ def logout():
 @app.route("/profile/delete", methods=("GET", "POST"))
 @login_required
 def delete_account():
+    if current_user.is_oauth_user == True:
+        abort(403)
     form = DeleteAccountForm()
     if form.validate_on_submit():
         if request.method == "POST" and bcrypt.check_password_hash(
@@ -405,7 +422,7 @@ def oauth2_callback(provider):
         password = "".join(random.choices(characters, k=20))
         hashed_password = bcrypt.generate_password_hash(password)
         user = User(
-            email=email, full_name=email.split("@")[0], password=hashed_password
+            email=email, full_name=email.split("@")[0], password=hashed_password,is_oauth_user = True
         )
         db.session.add(user)
         db.session.commit()
